fix deployments: staticfiles storage setting must be a valid import
